Scaffold-GS 推理显存优化与代码整理
## 优化效果

**测试场景**: MipNeRF360 Garden (2594×1680, 1.68M gaussians)

| 指标 | Baseline | Optimized | 提升 |
|------|----------|-----------|------|
| Generate 峰值 | 1,190.7 MB | 670.8 MB | -519.9 MB (-43.7%) |
| Rasterize 峰值 | 1,117.0 MB | 894.4 MB | -222.6 MB (-19.9%) |
| 渲染速度 (FP16) | 39.22 FPS | 60.97 FPS | +55.4% |
| 图像质量 | FP32 | PSNR 76.66 dB | 无损 |

## PyTorch 优化 (Generate 阶段)

- G1: Early Opacity Pruning - 提前过滤低 opacity 点
- G2: FP16 Mixed Precision - MLP 推理加速
- G3: Conditional Concatenation - 按需拼接特征
- G4: In-place Operations - 减少临时 tensor

**修改文件**: gaussian_renderer/__init__.py

## CUDA 优化 (Rasterize 阶段)

- R1: BinningState In-place Sorting - 排序复用 buffer
- R2: ImageState Ranges 修复 - 修正 254x 过度分配
- R2: uint16_t for n_contrib - 类型窄化节省显存
- R3: Inference Mode - 条件分配 backward-only buffers

**修改文件**:
- rasterizer_impl.h/cu - 核心内存分配逻辑
- forward.cu/h - 前向传播优化
- backward.cu/h - 后向传播接口更新
- forward_reweighted.cu, backward_reweighted.cu - 接口更新

## 代码整理

**测试脚本规范化**:
- render_profile_*.py → profile_memory_*.py
- validate_fp16_precision.py → profile_precision.py
- benchmark_fp16_vs_fp32.py → profile_speed.py

**测试结果组织**:
- profiling/results/memory/ - 显存测试数据
- profiling/results/speedup/ - 速度测试数据
- profiling/results/precision/ - 精度测试数据

**文档**:
- docs/OPTIMIZATION.md - 完整优化文档
- profiling/README.md - 测试脚本说明

**删除冗余文件**:
- test_optimization.py, test_speed_fp16.py (过时测试)
- test_fp16_anchor.png, test_fp32.png (失败实验)
- 中间版本测试数据

+++ submodules/diff-gaussian-rasterization/cuda_rasterizer/backward.h
@@ -30,7 +30,7 @@ namespace BACKWARD
 		const float4* conic_opacity,
 		const float* colors,
 		const float* final_Ts,
-		const uint32_t* n_contrib,
+		const uint16_t* n_contrib,
 		const float* dL_dpixels,
 		float3* dL_dmean2D,
 		float4* dL_dconic2D,
