+++ submodules/diff-gaussian-rasterization/cuda_rasterizer/forward.h
@@ -57,7 +57,7 @@ namespace FORWARD
 		const float* features,
 		const float4* conic_opacity,
 		float* final_T,
-		uint32_t* n_contrib,
+		uint16_t* n_contrib,
 		const float* bg_color,
 		float* out_color);
 	void filter_preprocess(int P, int M,
